LayoutTests/imported/w3c/web-platform-tests/css/css-color/parsing/color-mix-computed-expected.txt


PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%), hsl(30deg 30% 40%))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 25%, hsl(30deg 30% 40%))'
PASS Property color value 'color-mix(in hsl, 25% hsl(120deg 10% 20%), hsl(30deg 30% 40%))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%), 25% hsl(30deg 30% 40%))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%), hsl(30deg 30% 40%) 25%)'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 25%, hsl(30deg 30% 40%) 75%)'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 30%, hsl(30deg 30% 40%) 90%)'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 12.5%, hsl(30deg 30% 40%) 37.5%)'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 0%, hsl(30deg 30% 40%))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4), hsl(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20%) 25%, hsl(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hsl, 25% hsl(120deg 10% 20% / .4), hsl(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4), 25% hsl(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4), hsl(30deg 30% 40% / .8) 25%)'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4) 25%, hsl(30deg 30% 40% / .8) 75%)'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4) 30%, hsl(30deg 30% 40% / .8) 90%)'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4) 12.5%, hsl(30deg 30% 40% / .8) 37.5%)'
PASS Property color value 'color-mix(in hsl, hsl(120deg 10% 20% / .4) 0%, hsl(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hsl, hsl(40deg 50% 50%), hsl(60deg 50% 50%))'
PASS Property color value 'color-mix(in hsl, hsl(60deg 50% 50%), hsl(40deg 50% 50%))'
PASS Property color value 'color-mix(in hsl, hsl(50deg 50% 50%), hsl(330deg 50% 50%))'
PASS Property color value 'color-mix(in hsl, hsl(330deg 50% 50%), hsl(50deg 50% 50%))'
PASS Property color value 'color-mix(in hsl, hsl(20deg 50% 50%), hsl(320deg 50% 50%))'
PASS Property color value 'color-mix(in hsl, hsl(320deg 50% 50%), hsl(20deg 50% 50%))'
PASS Property color value 'color-mix(in hsl shorter hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))'
PASS Property color value 'color-mix(in hsl shorter hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))'
PASS Property color value 'color-mix(in hsl shorter hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))'
PASS Property color value 'color-mix(in hsl shorter hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))'
PASS Property color value 'color-mix(in hsl shorter hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))'
PASS Property color value 'color-mix(in hsl shorter hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))'
PASS Property color value 'color-mix(in hsl longer hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))'
PASS Property color value 'color-mix(in hsl longer hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))'
PASS Property color value 'color-mix(in hsl longer hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))'
PASS Property color value 'color-mix(in hsl longer hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))'
PASS Property color value 'color-mix(in hsl longer hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))'
PASS Property color value 'color-mix(in hsl longer hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))'
PASS Property color value 'color-mix(in hsl increasing hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))'
PASS Property color value 'color-mix(in hsl increasing hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))'
PASS Property color value 'color-mix(in hsl increasing hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))'
PASS Property color value 'color-mix(in hsl increasing hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))'
PASS Property color value 'color-mix(in hsl increasing hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))'
PASS Property color value 'color-mix(in hsl increasing hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))'
PASS Property color value 'color-mix(in hsl decreasing hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))'
PASS Property color value 'color-mix(in hsl decreasing hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))'
PASS Property color value 'color-mix(in hsl decreasing hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))'
PASS Property color value 'color-mix(in hsl decreasing hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))'
PASS Property color value 'color-mix(in hsl decreasing hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))'
PASS Property color value 'color-mix(in hsl decreasing hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))'
PASS Property color value 'color-mix(in hsl specified hue, hsl(40deg 50% 50%), hsl(60deg 50% 50%))'
PASS Property color value 'color-mix(in hsl specified hue, hsl(60deg 50% 50%), hsl(40deg 50% 50%))'
PASS Property color value 'color-mix(in hsl specified hue, hsl(50deg 50% 50%), hsl(330deg 50% 50%))'
PASS Property color value 'color-mix(in hsl specified hue, hsl(330deg 50% 50%), hsl(50deg 50% 50%))'
PASS Property color value 'color-mix(in hsl specified hue, hsl(20deg 50% 50%), hsl(320deg 50% 50%))'
PASS Property color value 'color-mix(in hsl specified hue, hsl(320deg 50% 50%), hsl(20deg 50% 50%))'
PASS Property color value 'color-mix(in hsl, hsl(none none none), hsl(none none none))'
PASS Property color value 'color-mix(in hsl, hsl(none none none), hsl(30deg 40% 80%))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 20% 40%), hsl(none none none))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 20% none), hsl(30deg 40% 60%))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 20% 40%), hsl(30deg 20% none))'
PASS Property color value 'color-mix(in hsl, hsl(none 20% 40%), hsl(30deg none 80%))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 40% 40% / none), hsl(0deg 40% 40%))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 40% 40% / none), hsl(0deg 40% 40% / 0.5))'
PASS Property color value 'color-mix(in hsl, hsl(120deg 40% 40% / none), hsl(0deg 40% 40% / none))'
PASS Property color value 'color-mix(in hsl, color(display-p3 0 1 0) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hsl, lab(100% 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hsl, lab(0% 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hsl, lch(100% 116 334) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hsl, lch(0% 116 334) 100%, rgb(0, 0, 0) 0%)'
FAIL Property color value 'color-mix(in hsl, oklab(100% 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' assert_equals: expected "rgb(255, 255, 255)" but got "rgb(255, 249, 255)"
PASS Property color value 'color-mix(in hsl, oklab(0% 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)'
FAIL Property color value 'color-mix(in hsl, oklch(100% 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' assert_equals: expected "rgb(255, 255, 255)" but got "rgb(255, 249, 255)"
PASS Property color value 'color-mix(in hsl, oklch(0% 0.399 336.3) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), hwb(30deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 25%, hwb(30deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, 25% hwb(120deg 10% 20%), hwb(30deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), 25% hwb(30deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), hwb(30deg 30% 40%) 25%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 25%, hwb(30deg 30% 40%) 75%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 30%, hwb(30deg 30% 40%) 90%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 12.5%, hwb(30deg 30% 40%) 37.5%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%) 0%, hwb(30deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4), hwb(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 25%, hwb(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hwb, 25% hwb(120deg 10% 20% / .4), hwb(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), 25% hwb(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4), hwb(30deg 30% 40% / .8) 25%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 25%, hwb(30deg 30% 40% / .8) 75%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 30%, hwb(30deg 30% 40% / .8) 90%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 12.5%, hwb(30deg 30% 40% / .8) 37.5%)'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / .4) 0%, hwb(30deg 30% 40% / .8))'
PASS Property color value 'color-mix(in hwb, hwb(40deg 30% 40%), hwb(60deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(60deg 30% 40%), hwb(40deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(50deg 30% 40%), hwb(330deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(330deg 30% 40%), hwb(50deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(20deg 30% 40%), hwb(320deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(320deg 30% 40%), hwb(20deg 30% 40%))'
PASS Property color value 'color-mix(in hwb shorter hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))'
PASS Property color value 'color-mix(in hwb shorter hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))'
PASS Property color value 'color-mix(in hwb shorter hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))'
PASS Property color value 'color-mix(in hwb shorter hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))'
PASS Property color value 'color-mix(in hwb shorter hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))'
PASS Property color value 'color-mix(in hwb shorter hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))'
PASS Property color value 'color-mix(in hwb longer hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))'
PASS Property color value 'color-mix(in hwb longer hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))'
PASS Property color value 'color-mix(in hwb longer hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))'
PASS Property color value 'color-mix(in hwb longer hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))'
PASS Property color value 'color-mix(in hwb longer hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))'
PASS Property color value 'color-mix(in hwb longer hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))'
PASS Property color value 'color-mix(in hwb increasing hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))'
PASS Property color value 'color-mix(in hwb increasing hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))'
PASS Property color value 'color-mix(in hwb increasing hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))'
PASS Property color value 'color-mix(in hwb increasing hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))'
PASS Property color value 'color-mix(in hwb increasing hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))'
PASS Property color value 'color-mix(in hwb increasing hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))'
PASS Property color value 'color-mix(in hwb decreasing hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))'
PASS Property color value 'color-mix(in hwb decreasing hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))'
PASS Property color value 'color-mix(in hwb decreasing hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))'
PASS Property color value 'color-mix(in hwb decreasing hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))'
PASS Property color value 'color-mix(in hwb decreasing hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))'
PASS Property color value 'color-mix(in hwb decreasing hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))'
PASS Property color value 'color-mix(in hwb specified hue, hwb(40deg 30% 40%), hwb(60deg 30% 40%))'
PASS Property color value 'color-mix(in hwb specified hue, hwb(60deg 30% 40%), hwb(40deg 30% 40%))'
PASS Property color value 'color-mix(in hwb specified hue, hwb(50deg 30% 40%), hwb(330deg 30% 40%))'
PASS Property color value 'color-mix(in hwb specified hue, hwb(330deg 30% 40%), hwb(50deg 30% 40%))'
PASS Property color value 'color-mix(in hwb specified hue, hwb(20deg 30% 40%), hwb(320deg 30% 40%))'
PASS Property color value 'color-mix(in hwb specified hue, hwb(320deg 30% 40%), hwb(20deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(none none none), hwb(none none none))'
PASS Property color value 'color-mix(in hwb, hwb(none none none), hwb(30deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), hwb(none none none))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% none), hwb(30deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20%), hwb(30deg 30% none))'
PASS Property color value 'color-mix(in hwb, hwb(none 10% 20%), hwb(30deg none 40%))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / none), hwb(30deg 30% 40%))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / none), hwb(30deg 30% 40% / 0.5))'
PASS Property color value 'color-mix(in hwb, hwb(120deg 10% 20% / none), hwb(30deg 30% 40% / none))'
PASS Property color value 'color-mix(in hwb, color(display-p3 0 1 0) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hwb, lab(100% 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hwb, lab(0% 104.3 -50.9) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hwb, lch(100% 116 334) 100%, rgb(0, 0, 0) 0%)'
PASS Property color value 'color-mix(in hwb, lch(0% 116 334) 100%, rgb(0, 0, 0) 0%)'
FAIL Property color value 'color-mix(in hwb, oklab(100% 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)' assert_equals: expected "rgb(255, 255, 255)" but got "rgb(255, 249, 255)"
PASS Property color value 'color-mix(in hwb, oklab(0% 0.365 -0.16) 100%, rgb(0, 0, 0) 0%)'
FAIL Property color value 'color-mix(in hwb, oklch(100% 0.399 336.3) 100%, rgb(0, 0, 0) 0%)' assert_equals: expected "rgb(255, 255, 255)" but got "rgb(255, 249, 255)"
PASS Property color value 'color-mix(in hwb, oklch(0% 0.399 336.3) 100%, rgb(0, 0, 0) 0%)'
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg), lch(50% 60 70deg))' assert_equals: expected "lch(30% 40 50)" but got "lch(30 40 50)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg) 25%, lch(50% 60 70deg))' assert_equals: expected "lch(40% 50 60)" but got "lch(40 50 60)"
FAIL Property color value 'color-mix(in lch, 25% lch(10% 20 30deg), lch(50% 60 70deg))' assert_equals: expected "lch(40% 50 60)" but got "lch(40 50 60)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg), 25% lch(50% 60 70deg))' assert_equals: expected "lch(20% 30 40)" but got "lch(20 30 40)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg), lch(50% 60 70deg) 25%)' assert_equals: expected "lch(20% 30 40)" but got "lch(20 30 40)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg) 25%, lch(50% 60 70deg) 75%)' assert_equals: expected "lch(40% 50 60)" but got "lch(40 50 60)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg) 30%, lch(50% 60 70deg) 90%)' assert_equals: expected "lch(40% 50 60)" but got "lch(40 50 60)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg) 12.5%, lch(50% 60 70deg) 37.5%)' assert_equals: expected "lch(40% 50 60 / 0.5)" but got "lch(40 50 60 / 0.5)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg) 0%, lch(50% 60 70deg))' assert_equals: expected "lch(50% 60 70)" but got "lch(50 60 70)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / .4), lch(50% 60 70deg / .8))' assert_equals: expected "lch(36.666664% 46.666664 50 / 0.6)" but got "lch(36.666664 46.666664 50 / 0.6)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / .4) 25%, lch(50% 60 70deg / .8))' assert_equals: expected "lch(44.285713% 54.285717 60 / 0.7)" but got "lch(44.285713 54.285717 60 / 0.7)"
FAIL Property color value 'color-mix(in lch, 25% lch(10% 20 30deg / .4), lch(50% 60 70deg / .8))' assert_equals: expected "lch(44.285713% 54.285717 60 / 0.7)" but got "lch(44.285713 54.285717 60 / 0.7)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / .4), 25% lch(50% 60 70deg / .8))' assert_equals: expected "lch(26% 36 40 / 0.5)" but got "lch(26 36 40 / 0.5)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / .4), lch(50% 60 70deg / .8) 25%)' assert_equals: expected "lch(26% 36 40 / 0.5)" but got "lch(26 36 40 / 0.5)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / .4) 25%, lch(50% 60 70deg / .8) 75%)' assert_equals: expected "lch(44.285713% 54.285717 60 / 0.7)" but got "lch(44.285713 54.285717 60 / 0.7)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / .4) 30%, lch(50% 60 70deg / .8) 90%)' assert_equals: expected "lch(44.285713% 54.285717 60 / 0.7)" but got "lch(44.285713 54.285717 60 / 0.7)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / .4) 12.5%, lch(50% 60 70deg / .8) 37.5%)' assert_equals: expected "lch(44.285713% 54.285717 60 / 0.35)" but got "lch(44.285713 54.285717 60 / 0.35)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / .4) 0%, lch(50% 60 70deg / .8))' assert_equals: expected "lch(50% 60 70 / 0.8)" but got "lch(50 60 70 / 0.8)"
FAIL Property color value 'color-mix(in lch, lch(100% 0 40deg), lch(100% 0 60deg))' assert_equals: expected "lch(100% 0 50)" but got "lch(100 0 50)"
FAIL Property color value 'color-mix(in lch, lch(100% 0 60deg), lch(100% 0 40deg))' assert_equals: expected "lch(100% 0 50)" but got "lch(100 0 50)"
FAIL Property color value 'color-mix(in lch, lch(100% 0 50deg), lch(100% 0 330deg))' assert_equals: expected "lch(100% 0 10)" but got "lch(100 0 10)"
FAIL Property color value 'color-mix(in lch, lch(100% 0 330deg), lch(100% 0 50deg))' assert_equals: expected "lch(100% 0 10)" but got "lch(100 0 10)"
FAIL Property color value 'color-mix(in lch, lch(100% 0 20deg), lch(100% 0 320deg))' assert_equals: expected "lch(100% 0 350)" but got "lch(100 0 350)"
FAIL Property color value 'color-mix(in lch, lch(100% 0 320deg), lch(100% 0 20deg))' assert_equals: expected "lch(100% 0 350)" but got "lch(100 0 350)"
FAIL Property color value 'color-mix(in lch shorter hue, lch(100% 0 40deg), lch(100% 0 60deg))' assert_equals: expected "lch(100% 0 50)" but got "lch(100 0 50)"
FAIL Property color value 'color-mix(in lch shorter hue, lch(100% 0 60deg), lch(100% 0 40deg))' assert_equals: expected "lch(100% 0 50)" but got "lch(100 0 50)"
FAIL Property color value 'color-mix(in lch shorter hue, lch(100% 0 50deg), lch(100% 0 330deg))' assert_equals: expected "lch(100% 0 10)" but got "lch(100 0 10)"
FAIL Property color value 'color-mix(in lch shorter hue, lch(100% 0 330deg), lch(100% 0 50deg))' assert_equals: expected "lch(100% 0 10)" but got "lch(100 0 10)"
FAIL Property color value 'color-mix(in lch shorter hue, lch(100% 0 20deg), lch(100% 0 320deg))' assert_equals: expected "lch(100% 0 350)" but got "lch(100 0 350)"
FAIL Property color value 'color-mix(in lch shorter hue, lch(100% 0 320deg), lch(100% 0 20deg))' assert_equals: expected "lch(100% 0 350)" but got "lch(100 0 350)"
FAIL Property color value 'color-mix(in lch longer hue, lch(100% 0 40deg), lch(100% 0 60deg))' assert_equals: expected "lch(100% 0 230)" but got "lch(100 0 230)"
FAIL Property color value 'color-mix(in lch longer hue, lch(100% 0 60deg), lch(100% 0 40deg))' assert_equals: expected "lch(100% 0 230)" but got "lch(100 0 230)"
FAIL Property color value 'color-mix(in lch longer hue, lch(100% 0 50deg), lch(100% 0 330deg))' assert_equals: expected "lch(100% 0 190)" but got "lch(100 0 190)"
FAIL Property color value 'color-mix(in lch longer hue, lch(100% 0 330deg), lch(100% 0 50deg))' assert_equals: expected "lch(100% 0 190)" but got "lch(100 0 190)"
FAIL Property color value 'color-mix(in lch longer hue, lch(100% 0 20deg), lch(100% 0 320deg))' assert_equals: expected "lch(100% 0 170)" but got "lch(100 0 170)"
FAIL Property color value 'color-mix(in lch longer hue, lch(100% 0 320deg), lch(100% 0 20deg))' assert_equals: expected "lch(100% 0 170)" but got "lch(100 0 170)"
FAIL Property color value 'color-mix(in lch increasing hue, lch(100% 0 40deg), lch(100% 0 60deg))' assert_equals: expected "lch(100% 0 50)" but got "lch(100 0 50)"
FAIL Property color value 'color-mix(in lch increasing hue, lch(100% 0 60deg), lch(100% 0 40deg))' assert_equals: expected "lch(100% 0 230)" but got "lch(100 0 230)"
FAIL Property color value 'color-mix(in lch increasing hue, lch(100% 0 50deg), lch(100% 0 330deg))' assert_equals: expected "lch(100% 0 190)" but got "lch(100 0 190)"
FAIL Property color value 'color-mix(in lch increasing hue, lch(100% 0 330deg), lch(100% 0 50deg))' assert_equals: expected "lch(100% 0 10)" but got "lch(100 0 10)"
FAIL Property color value 'color-mix(in lch increasing hue, lch(100% 0 20deg), lch(100% 0 320deg))' assert_equals: expected "lch(100% 0 170)" but got "lch(100 0 170)"
FAIL Property color value 'color-mix(in lch increasing hue, lch(100% 0 320deg), lch(100% 0 20deg))' assert_equals: expected "lch(100% 0 350)" but got "lch(100 0 350)"
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100% 0 40deg), lch(100% 0 60deg))' assert_equals: expected "lch(100% 0 230)" but got "lch(100 0 230)"
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100% 0 60deg), lch(100% 0 40deg))' assert_equals: expected "lch(100% 0 50)" but got "lch(100 0 50)"
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100% 0 50deg), lch(100% 0 330deg))' assert_equals: expected "lch(100% 0 10)" but got "lch(100 0 10)"
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100% 0 330deg), lch(100% 0 50deg))' assert_equals: expected "lch(100% 0 190)" but got "lch(100 0 190)"
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100% 0 20deg), lch(100% 0 320deg))' assert_equals: expected "lch(100% 0 350)" but got "lch(100 0 350)"
FAIL Property color value 'color-mix(in lch decreasing hue, lch(100% 0 320deg), lch(100% 0 20deg))' assert_equals: expected "lch(100% 0 170)" but got "lch(100 0 170)"
FAIL Property color value 'color-mix(in lch specified hue, lch(100% 0 40deg), lch(100% 0 60deg))' assert_equals: expected "lch(100% 0 50)" but got "lch(100 0 50)"
FAIL Property color value 'color-mix(in lch specified hue, lch(100% 0 60deg), lch(100% 0 40deg))' assert_equals: expected "lch(100% 0 50)" but got "lch(100 0 50)"
FAIL Property color value 'color-mix(in lch specified hue, lch(100% 0 50deg), lch(100% 0 330deg))' assert_equals: expected "lch(100% 0 190)" but got "lch(100 0 190)"
FAIL Property color value 'color-mix(in lch specified hue, lch(100% 0 330deg), lch(100% 0 50deg))' assert_equals: expected "lch(100% 0 190)" but got "lch(100 0 190)"
FAIL Property color value 'color-mix(in lch specified hue, lch(100% 0 20deg), lch(100% 0 320deg))' assert_equals: expected "lch(100% 0 170)" but got "lch(100 0 170)"
FAIL Property color value 'color-mix(in lch specified hue, lch(100% 0 320deg), lch(100% 0 20deg))' assert_equals: expected "lch(100% 0 170)" but got "lch(100 0 170)"
PASS Property color value 'color-mix(in lch, lch(none none none), lch(none none none))'
FAIL Property color value 'color-mix(in lch, lch(none none none), lch(50% 60 70deg))' assert_equals: expected "lch(50% 60 70)" but got "lch(50 60 70)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg), lch(none none none))' assert_equals: expected "lch(10% 20 30)" but got "lch(10 20 30)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 none), lch(50% 60 70deg))' assert_equals: expected "lch(30% 40 70)" but got "lch(30 40 70)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg), lch(50% 60 none))' assert_equals: expected "lch(30% 40 30)" but got "lch(30 40 30)"
FAIL Property color value 'color-mix(in lch, lch(none 20 30deg), lch(50% none 70deg))' assert_equals: expected "lch(50% 20 50)" but got "lch(50 20 50)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / none), lch(50% 60 70deg))' assert_equals: expected "lch(30% 40 50)" but got "lch(30 40 50)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / none), lch(50% 60 70deg / 0.5))' assert_equals: expected "lch(30% 40 50 / 0.5)" but got "lch(30 40 50 / 0.5)"
FAIL Property color value 'color-mix(in lch, lch(10% 20 30deg / none), lch(50% 60 70deg / none))' assert_equals: expected "lch(30% 40 50 / none)" but got "lch(30 40 50 / none)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg), oklch(50% 60 70deg))' assert_equals: expected "oklch(30% 40 50)" but got "oklch(0.3 40 50)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg) 25%, oklch(50% 60 70deg))' assert_equals: expected "oklch(40% 50 60)" but got "oklch(0.4 50 60)"
FAIL Property color value 'color-mix(in oklch, 25% oklch(10% 20 30deg), oklch(50% 60 70deg))' assert_equals: expected "oklch(40% 50 60)" but got "oklch(0.4 50 60)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg), 25% oklch(50% 60 70deg))' assert_equals: expected "oklch(20% 30 40)" but got "oklch(0.2 30 40)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg), oklch(50% 60 70deg) 25%)' assert_equals: expected "oklch(20% 30 40)" but got "oklch(0.2 30 40)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg) 25%, oklch(50% 60 70deg) 75%)' assert_equals: expected "oklch(40% 50 60)" but got "oklch(0.4 50 60)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg) 30%, oklch(50% 60 70deg) 90%)' assert_equals: expected "oklch(40% 50 60)" but got "oklch(0.4 50 60)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg) 12.5%, oklch(50% 60 70deg) 37.5%)' assert_equals: expected "oklch(40% 50 60 / 0.5)" but got "oklch(0.4 50 60 / 0.5)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg) 0%, oklch(50% 60 70deg))' assert_equals: expected "oklch(50% 60 70)" but got "oklch(0.5 60 70)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / .4), oklch(50% 60 70deg / .8))' assert_equals: expected "oklch(36.666664% 46.666664 50 / 0.6)" but got "oklch(0.36666664 46.666664 50 / 0.6)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / .4) 25%, oklch(50% 60 70deg / .8))' assert_equals: expected "oklch(44.285713% 54.285717 60 / 0.7)" but got "oklch(0.44285715 54.285717 60 / 0.7)"
FAIL Property color value 'color-mix(in oklch, 25% oklch(10% 20 30deg / .4), oklch(50% 60 70deg / .8))' assert_equals: expected "oklch(44.285713% 54.285717 60 / 0.7)" but got "oklch(0.44285715 54.285717 60 / 0.7)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / .4), 25% oklch(50% 60 70deg / .8))' assert_equals: expected "oklch(26% 36 40 / 0.5)" but got "oklch(0.26000002 36 40 / 0.5)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / .4), oklch(50% 60 70deg / .8) 25%)' assert_equals: expected "oklch(26% 36 40 / 0.5)" but got "oklch(0.26000002 36 40 / 0.5)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / .4) 25%, oklch(50% 60 70deg / .8) 75%)' assert_equals: expected "oklch(44.285713% 54.285717 60 / 0.7)" but got "oklch(0.44285715 54.285717 60 / 0.7)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / .4) 30%, oklch(50% 60 70deg / .8) 90%)' assert_equals: expected "oklch(44.285713% 54.285717 60 / 0.7)" but got "oklch(0.44285715 54.285717 60 / 0.7)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / .4) 12.5%, oklch(50% 60 70deg / .8) 37.5%)' assert_equals: expected "oklch(44.285713% 54.285717 60 / 0.35)" but got "oklch(0.44285715 54.285717 60 / 0.35)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / .4) 0%, oklch(50% 60 70deg / .8))' assert_equals: expected "oklch(50% 60 70 / 0.8)" but got "oklch(0.5 60 70 / 0.8)"
FAIL Property color value 'color-mix(in oklch, oklch(100% 0 40deg), oklch(100% 0 60deg))' assert_equals: expected "oklch(100% 0 50)" but got "oklch(1 0 50)"
FAIL Property color value 'color-mix(in oklch, oklch(100% 0 60deg), oklch(100% 0 40deg))' assert_equals: expected "oklch(100% 0 50)" but got "oklch(1 0 50)"
FAIL Property color value 'color-mix(in oklch, oklch(100% 0 50deg), oklch(100% 0 330deg))' assert_equals: expected "oklch(100% 0 10)" but got "oklch(1 0 10)"
FAIL Property color value 'color-mix(in oklch, oklch(100% 0 330deg), oklch(100% 0 50deg))' assert_equals: expected "oklch(100% 0 10)" but got "oklch(1 0 10)"
FAIL Property color value 'color-mix(in oklch, oklch(100% 0 20deg), oklch(100% 0 320deg))' assert_equals: expected "oklch(100% 0 350)" but got "oklch(1 0 350)"
FAIL Property color value 'color-mix(in oklch, oklch(100% 0 320deg), oklch(100% 0 20deg))' assert_equals: expected "oklch(100% 0 350)" but got "oklch(1 0 350)"
FAIL Property color value 'color-mix(in oklch shorter hue, oklch(100% 0 40deg), oklch(100% 0 60deg))' assert_equals: expected "oklch(100% 0 50)" but got "oklch(1 0 50)"
FAIL Property color value 'color-mix(in oklch shorter hue, oklch(100% 0 60deg), oklch(100% 0 40deg))' assert_equals: expected "oklch(100% 0 50)" but got "oklch(1 0 50)"
FAIL Property color value 'color-mix(in oklch shorter hue, oklch(100% 0 50deg), oklch(100% 0 330deg))' assert_equals: expected "oklch(100% 0 10)" but got "oklch(1 0 10)"
FAIL Property color value 'color-mix(in oklch shorter hue, oklch(100% 0 330deg), oklch(100% 0 50deg))' assert_equals: expected "oklch(100% 0 10)" but got "oklch(1 0 10)"
FAIL Property color value 'color-mix(in oklch shorter hue, oklch(100% 0 20deg), oklch(100% 0 320deg))' assert_equals: expected "oklch(100% 0 350)" but got "oklch(1 0 350)"
FAIL Property color value 'color-mix(in oklch shorter hue, oklch(100% 0 320deg), oklch(100% 0 20deg))' assert_equals: expected "oklch(100% 0 350)" but got "oklch(1 0 350)"
FAIL Property color value 'color-mix(in oklch longer hue, oklch(100% 0 40deg), oklch(100% 0 60deg))' assert_equals: expected "oklch(100% 0 230)" but got "oklch(1 0 230)"
FAIL Property color value 'color-mix(in oklch longer hue, oklch(100% 0 60deg), oklch(100% 0 40deg))' assert_equals: expected "oklch(100% 0 230)" but got "oklch(1 0 230)"
FAIL Property color value 'color-mix(in oklch longer hue, oklch(100% 0 50deg), oklch(100% 0 330deg))' assert_equals: expected "oklch(100% 0 190)" but got "oklch(1 0 190)"
FAIL Property color value 'color-mix(in oklch longer hue, oklch(100% 0 330deg), oklch(100% 0 50deg))' assert_equals: expected "oklch(100% 0 190)" but got "oklch(1 0 190)"
FAIL Property color value 'color-mix(in oklch longer hue, oklch(100% 0 20deg), oklch(100% 0 320deg))' assert_equals: expected "oklch(100% 0 170)" but got "oklch(1 0 170)"
FAIL Property color value 'color-mix(in oklch longer hue, oklch(100% 0 320deg), oklch(100% 0 20deg))' assert_equals: expected "oklch(100% 0 170)" but got "oklch(1 0 170)"
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100% 0 40deg), oklch(100% 0 60deg))' assert_equals: expected "oklch(100% 0 50)" but got "oklch(1 0 50)"
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100% 0 60deg), oklch(100% 0 40deg))' assert_equals: expected "oklch(100% 0 230)" but got "oklch(1 0 230)"
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100% 0 50deg), oklch(100% 0 330deg))' assert_equals: expected "oklch(100% 0 190)" but got "oklch(1 0 190)"
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100% 0 330deg), oklch(100% 0 50deg))' assert_equals: expected "oklch(100% 0 10)" but got "oklch(1 0 10)"
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100% 0 20deg), oklch(100% 0 320deg))' assert_equals: expected "oklch(100% 0 170)" but got "oklch(1 0 170)"
FAIL Property color value 'color-mix(in oklch increasing hue, oklch(100% 0 320deg), oklch(100% 0 20deg))' assert_equals: expected "oklch(100% 0 350)" but got "oklch(1 0 350)"
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100% 0 40deg), oklch(100% 0 60deg))' assert_equals: expected "oklch(100% 0 230)" but got "oklch(1 0 230)"
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100% 0 60deg), oklch(100% 0 40deg))' assert_equals: expected "oklch(100% 0 50)" but got "oklch(1 0 50)"
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100% 0 50deg), oklch(100% 0 330deg))' assert_equals: expected "oklch(100% 0 10)" but got "oklch(1 0 10)"
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100% 0 330deg), oklch(100% 0 50deg))' assert_equals: expected "oklch(100% 0 190)" but got "oklch(1 0 190)"
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100% 0 20deg), oklch(100% 0 320deg))' assert_equals: expected "oklch(100% 0 350)" but got "oklch(1 0 350)"
FAIL Property color value 'color-mix(in oklch decreasing hue, oklch(100% 0 320deg), oklch(100% 0 20deg))' assert_equals: expected "oklch(100% 0 170)" but got "oklch(1 0 170)"
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100% 0 40deg), oklch(100% 0 60deg))' assert_equals: expected "oklch(100% 0 50)" but got "oklch(1 0 50)"
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100% 0 60deg), oklch(100% 0 40deg))' assert_equals: expected "oklch(100% 0 50)" but got "oklch(1 0 50)"
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100% 0 50deg), oklch(100% 0 330deg))' assert_equals: expected "oklch(100% 0 190)" but got "oklch(1 0 190)"
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100% 0 330deg), oklch(100% 0 50deg))' assert_equals: expected "oklch(100% 0 190)" but got "oklch(1 0 190)"
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100% 0 20deg), oklch(100% 0 320deg))' assert_equals: expected "oklch(100% 0 170)" but got "oklch(1 0 170)"
FAIL Property color value 'color-mix(in oklch specified hue, oklch(100% 0 320deg), oklch(100% 0 20deg))' assert_equals: expected "oklch(100% 0 170)" but got "oklch(1 0 170)"
PASS Property color value 'color-mix(in oklch, oklch(none none none), oklch(none none none))'
FAIL Property color value 'color-mix(in oklch, oklch(none none none), oklch(50% 60 70deg))' assert_equals: expected "oklch(50% 60 70)" but got "oklch(0.5 60 70)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg), oklch(none none none))' assert_equals: expected "oklch(10% 20 30)" but got "oklch(0.1 20 30)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 none), oklch(50% 60 70deg))' assert_equals: expected "oklch(30% 40 70)" but got "oklch(0.3 40 70)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg), oklch(50% 60 none))' assert_equals: expected "oklch(30% 40 30)" but got "oklch(0.3 40 30)"
FAIL Property color value 'color-mix(in oklch, oklch(none 20 30deg), oklch(50% none 70deg))' assert_equals: expected "oklch(50% 20 50)" but got "oklch(0.5 20 50)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / none), oklch(50% 60 70deg))' assert_equals: expected "oklch(30% 40 50)" but got "oklch(0.3 40 50)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / none), oklch(50% 60 70deg / 0.5))' assert_equals: expected "oklch(30% 40 50 / 0.5)" but got "oklch(0.3 40 50 / 0.5)"
FAIL Property color value 'color-mix(in oklch, oklch(10% 20 30deg / none), oklch(50% 60 70deg / none))' assert_equals: expected "oklch(30% 40 50 / none)" but got "oklch(0.3 40 50 / none)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30), lab(50% 60 70))' assert_equals: expected "lab(30% 40 50)" but got "lab(30 40 50)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30) 25%, lab(50% 60 70))' assert_equals: expected "lab(40% 50 60)" but got "lab(40 50 60)"
FAIL Property color value 'color-mix(in lab, 25% lab(10% 20 30), lab(50% 60 70))' assert_equals: expected "lab(40% 50 60)" but got "lab(40 50 60)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30), 25% lab(50% 60 70))' assert_equals: expected "lab(20% 30 40)" but got "lab(20 30 40)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30), lab(50% 60 70) 25%)' assert_equals: expected "lab(20% 30 40)" but got "lab(20 30 40)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30) 25%, lab(50% 60 70) 75%)' assert_equals: expected "lab(40% 50 60)" but got "lab(40 50 60)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30) 30%, lab(50% 60 70) 90%)' assert_equals: expected "lab(40% 50 60)" but got "lab(40 50 60)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30) 12.5%, lab(50% 60 70) 37.5%)' assert_equals: expected "lab(40% 50 60 / 0.5)" but got "lab(40 50 60 / 0.5)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30) 0%, lab(50% 60 70))' assert_equals: expected "lab(50% 60 70)" but got "lab(50 60 70)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / .4), lab(50% 60 70 / .8))' assert_equals: expected "lab(36.666664% 46.666664 56.666664 / 0.6)" but got "lab(36.666664 46.666664 56.666664 / 0.6)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / .4) 25%, lab(50% 60 70 / .8))' assert_equals: expected "lab(44.285713% 54.285717 64.28571 / 0.7)" but got "lab(44.285713 54.285717 64.28571 / 0.7)"
FAIL Property color value 'color-mix(in lab, 25% lab(10% 20 30 / .4), lab(50% 60 70 / .8))' assert_equals: expected "lab(44.285713% 54.285717 64.28571 / 0.7)" but got "lab(44.285713 54.285717 64.28571 / 0.7)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / .4), 25% lab(50% 60 70 / .8))' assert_equals: expected "lab(26% 36 46 / 0.5)" but got "lab(26 36 46 / 0.5)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / .4), lab(50% 60 70 / .8) 25%)' assert_equals: expected "lab(26% 36 46 / 0.5)" but got "lab(26 36 46 / 0.5)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / .4) 25%, lab(50% 60 70 / .8) 75%)' assert_equals: expected "lab(44.285713% 54.285717 64.28571 / 0.7)" but got "lab(44.285713 54.285717 64.28571 / 0.7)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / .4) 30%, lab(50% 60 70 / .8) 90%)' assert_equals: expected "lab(44.285713% 54.285717 64.28571 / 0.7)" but got "lab(44.285713 54.285717 64.28571 / 0.7)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / .4) 12.5%, lab(50% 60 70 / .8) 37.5%)' assert_equals: expected "lab(44.285713% 54.285717 64.28571 / 0.35)" but got "lab(44.285713 54.285717 64.28571 / 0.35)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / .4) 0%, lab(50% 60 70 / .8))' assert_equals: expected "lab(50% 60 70 / 0.8)" but got "lab(50 60 70 / 0.8)"
PASS Property color value 'color-mix(in lab, lab(none none none), lab(none none none))'
FAIL Property color value 'color-mix(in lab, lab(none none none), lab(50% 60 70))' assert_equals: expected "lab(50% 60 70)" but got "lab(50 60 70)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30), lab(none none none))' assert_equals: expected "lab(10% 20 30)" but got "lab(10 20 30)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 none), lab(50% 60 70))' assert_equals: expected "lab(30% 40 70)" but got "lab(30 40 70)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30), lab(50% 60 none))' assert_equals: expected "lab(30% 40 30)" but got "lab(30 40 30)"
FAIL Property color value 'color-mix(in lab, lab(none 20 30), lab(50% none 70))' assert_equals: expected "lab(50% 20 50)" but got "lab(50 20 50)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / none), lab(50% 60 70))' assert_equals: expected "lab(30% 40 50)" but got "lab(30 40 50)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / none), lab(50% 60 70 / 0.5))' assert_equals: expected "lab(30% 40 50 / 0.5)" but got "lab(30 40 50 / 0.5)"
FAIL Property color value 'color-mix(in lab, lab(10% 20 30 / none), lab(50% 60 70 / none))' assert_equals: expected "lab(30% 40 50 / none)" but got "lab(30 40 50 / none)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30), oklab(50% 60 70))' assert_equals: expected "oklab(30% 40 50)" but got "oklab(0.3 40 50)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30) 25%, oklab(50% 60 70))' assert_equals: expected "oklab(40% 50 60)" but got "oklab(0.4 50 60)"
FAIL Property color value 'color-mix(in oklab, 25% oklab(10% 20 30), oklab(50% 60 70))' assert_equals: expected "oklab(40% 50 60)" but got "oklab(0.4 50 60)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30), 25% oklab(50% 60 70))' assert_equals: expected "oklab(20% 30 40)" but got "oklab(0.2 30 40)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30), oklab(50% 60 70) 25%)' assert_equals: expected "oklab(20% 30 40)" but got "oklab(0.2 30 40)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30) 25%, oklab(50% 60 70) 75%)' assert_equals: expected "oklab(40% 50 60)" but got "oklab(0.4 50 60)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30) 30%, oklab(50% 60 70) 90%)' assert_equals: expected "oklab(40% 50 60)" but got "oklab(0.4 50 60)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30) 12.5%, oklab(50% 60 70) 37.5%)' assert_equals: expected "oklab(40% 50 60 / 0.5)" but got "oklab(0.4 50 60 / 0.5)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30) 0%, oklab(50% 60 70))' assert_equals: expected "oklab(50% 60 70)" but got "oklab(0.5 60 70)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / .4), oklab(50% 60 70 / .8))' assert_equals: expected "oklab(36.666664% 46.666664 56.666664 / 0.6)" but got "oklab(0.36666664 46.666664 56.666664 / 0.6)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / .4) 25%, oklab(50% 60 70 / .8))' assert_equals: expected "oklab(44.285713% 54.285717 64.28571 / 0.7)" but got "oklab(0.44285715 54.285717 64.28571 / 0.7)"
FAIL Property color value 'color-mix(in oklab, 25% oklab(10% 20 30 / .4), oklab(50% 60 70 / .8))' assert_equals: expected "oklab(44.285713% 54.285717 64.28571 / 0.7)" but got "oklab(0.44285715 54.285717 64.28571 / 0.7)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / .4), 25% oklab(50% 60 70 / .8))' assert_equals: expected "oklab(26% 36 46 / 0.5)" but got "oklab(0.26000002 36 46 / 0.5)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / .4), oklab(50% 60 70 / .8) 25%)' assert_equals: expected "oklab(26% 36 46 / 0.5)" but got "oklab(0.26000002 36 46 / 0.5)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / .4) 25%, oklab(50% 60 70 / .8) 75%)' assert_equals: expected "oklab(44.285713% 54.285717 64.28571 / 0.7)" but got "oklab(0.44285715 54.285717 64.28571 / 0.7)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / .4) 30%, oklab(50% 60 70 / .8) 90%)' assert_equals: expected "oklab(44.285713% 54.285717 64.28571 / 0.7)" but got "oklab(0.44285715 54.285717 64.28571 / 0.7)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / .4) 12.5%, oklab(50% 60 70 / .8) 37.5%)' assert_equals: expected "oklab(44.285713% 54.285717 64.28571 / 0.35)" but got "oklab(0.44285715 54.285717 64.28571 / 0.35)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / .4) 0%, oklab(50% 60 70 / .8))' assert_equals: expected "oklab(50% 60 70 / 0.8)" but got "oklab(0.5 60 70 / 0.8)"
PASS Property color value 'color-mix(in oklab, oklab(none none none), oklab(none none none))'
FAIL Property color value 'color-mix(in oklab, oklab(none none none), oklab(50% 60 70))' assert_equals: expected "oklab(50% 60 70)" but got "oklab(0.5 60 70)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30), oklab(none none none))' assert_equals: expected "oklab(10% 20 30)" but got "oklab(0.1 20 30)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 none), oklab(50% 60 70))' assert_equals: expected "oklab(30% 40 70)" but got "oklab(0.3 40 70)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30), oklab(50% 60 none))' assert_equals: expected "oklab(30% 40 30)" but got "oklab(0.3 40 30)"
FAIL Property color value 'color-mix(in oklab, oklab(none 20 30), oklab(50% none 70))' assert_equals: expected "oklab(50% 20 50)" but got "oklab(0.5 20 50)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / none), oklab(50% 60 70))' assert_equals: expected "oklab(30% 40 50)" but got "oklab(0.3 40 50)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / none), oklab(50% 60 70 / 0.5))' assert_equals: expected "oklab(30% 40 50 / 0.5)" but got "oklab(0.3 40 50 / 0.5)"
FAIL Property color value 'color-mix(in oklab, oklab(10% 20 30 / none), oklab(50% 60 70 / none))' assert_equals: expected "oklab(30% 40 50 / none)" but got "oklab(0.3 40 50 / none)"
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 25%, color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, 25% color(srgb .1 .2 .3), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), color(srgb .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), 25% color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 25%, color(srgb .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 30%, color(srgb .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 12.5%, color(srgb .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3) 0%, color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .5), color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 25%, color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, 25% color(srgb .1 .2 .3 / .4), color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4), color(srgb .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4), 25% color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 25%, color(srgb .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 30%, color(srgb .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 12.5%, color(srgb .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / .4) 0%, color(srgb .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb, color(srgb 2 3 4 / 5), color(srgb 4 6 8 / 10))'
PASS Property color value 'color-mix(in srgb, color(srgb -2 -3 -4), color(srgb -4 -6 -8))'
PASS Property color value 'color-mix(in srgb, color(srgb -2 -3 -4 / -5), color(srgb -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in srgb, color(srgb none none none), color(srgb none none none))'
PASS Property color value 'color-mix(in srgb, color(srgb none none none), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), color(srgb none none none))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 none), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3), color(srgb .5 .6 none))'
PASS Property color value 'color-mix(in srgb, color(srgb none .2 .3), color(srgb .5 none .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / none), color(srgb .5 .6 .7))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / none), color(srgb .5 .6 .7 / 0.5))'
PASS Property color value 'color-mix(in srgb, color(srgb .1 .2 .3 / none), color(srgb .5 .6 .7 / none))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 25%, color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, 25% color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), 25% color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 25%, color(srgb-linear .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 30%, color(srgb-linear .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 12.5%, color(srgb-linear .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3) 0%, color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .5), color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 25%, color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, 25% color(srgb-linear .1 .2 .3 / .4), color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4), color(srgb-linear .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4), 25% color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 25%, color(srgb-linear .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 30%, color(srgb-linear .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 12.5%, color(srgb-linear .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / .4) 0%, color(srgb-linear .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear 2 3 4 / 5), color(srgb-linear 4 6 8 / 10))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear -2 -3 -4), color(srgb-linear -4 -6 -8))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear -2 -3 -4 / -5), color(srgb-linear -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear none none none), color(srgb-linear none none none))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear none none none), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), color(srgb-linear none none none))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 none), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3), color(srgb-linear .5 .6 none))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear none .2 .3), color(srgb-linear .5 none .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / none), color(srgb-linear .5 .6 .7))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / none), color(srgb-linear .5 .6 .7 / 0.5))'
PASS Property color value 'color-mix(in srgb-linear, color(srgb-linear .1 .2 .3 / none), color(srgb-linear .5 .6 .7 / none))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 25%, color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, 25% color(xyz .1 .2 .3), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), color(xyz .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), 25% color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 25%, color(xyz .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 30%, color(xyz .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 12.5%, color(xyz .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3) 0%, color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .5), color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 25%, color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, 25% color(xyz .1 .2 .3 / .4), color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4), color(xyz .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4), 25% color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 25%, color(xyz .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 30%, color(xyz .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 12.5%, color(xyz .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / .4) 0%, color(xyz .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz, color(xyz 2 3 4 / 5), color(xyz 4 6 8 / 10))'
PASS Property color value 'color-mix(in xyz, color(xyz -2 -3 -4), color(xyz -4 -6 -8))'
PASS Property color value 'color-mix(in xyz, color(xyz -2 -3 -4 / -5), color(xyz -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in xyz, color(xyz none none none), color(xyz none none none))'
PASS Property color value 'color-mix(in xyz, color(xyz none none none), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), color(xyz none none none))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 none), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3), color(xyz .5 .6 none))'
PASS Property color value 'color-mix(in xyz, color(xyz none .2 .3), color(xyz .5 none .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / none), color(xyz .5 .6 .7))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / none), color(xyz .5 .6 .7 / 0.5))'
PASS Property color value 'color-mix(in xyz, color(xyz .1 .2 .3 / none), color(xyz .5 .6 .7 / none))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 25%, color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, 25% color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), 25% color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 25%, color(xyz-d50 .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 30%, color(xyz-d50 .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 12.5%, color(xyz-d50 .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3) 0%, color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .5), color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 25%, color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, 25% color(xyz-d50 .1 .2 .3 / .4), color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4), color(xyz-d50 .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4), 25% color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 25%, color(xyz-d50 .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 30%, color(xyz-d50 .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 12.5%, color(xyz-d50 .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / .4) 0%, color(xyz-d50 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 2 3 4 / 5), color(xyz-d50 4 6 8 / 10))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 -2 -3 -4), color(xyz-d50 -4 -6 -8))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 -2 -3 -4 / -5), color(xyz-d50 -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 none none none), color(xyz-d50 none none none))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 none none none), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), color(xyz-d50 none none none))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 none), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3), color(xyz-d50 .5 .6 none))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 none .2 .3), color(xyz-d50 .5 none .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / none), color(xyz-d50 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / none), color(xyz-d50 .5 .6 .7 / 0.5))'
PASS Property color value 'color-mix(in xyz-d50, color(xyz-d50 .1 .2 .3 / none), color(xyz-d50 .5 .6 .7 / none))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 25%, color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, 25% color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 .7) 25%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), 25% color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 25%, color(xyz-d65 .5 .6 .7) 75%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 30%, color(xyz-d65 .5 .6 .7) 90%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 12.5%, color(xyz-d65 .5 .6 .7) 37.5%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3) 0%, color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .5), color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 25%, color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, 25% color(xyz-d65 .1 .2 .3 / .4), color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4), color(xyz-d65 .5 .6 .7 / .8) 25%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4), 25% color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 25%, color(xyz-d65 .5 .6 .7 / .8) 75%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 30%, color(xyz-d65 .5 .6 .7 / .8) 90%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 12.5%, color(xyz-d65 .5 .6 .7 / .8) 37.5%)'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / .4) 0%, color(xyz-d65 .5 .6 .7 / .8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 2 3 4 / 5), color(xyz-d65 4 6 8 / 10))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 -2 -3 -4), color(xyz-d65 -4 -6 -8))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 -2 -3 -4 / -5), color(xyz-d65 -4 -6 -8 / -10))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 none none none), color(xyz-d65 none none none))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 none none none), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), color(xyz-d65 none none none))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 none), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3), color(xyz-d65 .5 .6 none))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 none .2 .3), color(xyz-d65 .5 none .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / none), color(xyz-d65 .5 .6 .7))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / none), color(xyz-d65 .5 .6 .7 / 0.5))'
PASS Property color value 'color-mix(in xyz-d65, color(xyz-d65 .1 .2 .3 / none), color(xyz-d65 .5 .6 .7 / none))'
